requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
pillow==10.3.0
transformers==4.41.2
requests==2.31.0
streamlit==1.35.0
torch==2.7.1
python-multipart==0.0.5


